Move spread.py tests into spread_test.py
